Fix engine.Dot: do not convert x and y to uint16 for json marshaling

+++ engine/dot.go
@@ -17,7 +17,7 @@ func (d1 Dot) Equals(d2 Dot) bool {
 
 // Implementing json.Marshaler interface
 func (d Dot) MarshalJSON() ([]byte, error) {
-	return json.Marshal([]uint16{uint16(d.X), uint16(d.Y)})
+	return json.Marshal([]uint8{d.X, d.Y})
 }
 
 func (d Dot) String() string {
